requirements.txt
python-dotenv
scrapy
beautifulsoup4
pandas
pymongo
sentence_transformers
streamlit
openai
fastapi
spacy 
scikit-learn
pyvi
numpy==1.26.4
uvicorn
motor